Change Distance type from size_t to double.

+++ C++ for C Programmers, Part A/Module2/homework/dijkstra.cpp
@@ -52,7 +52,7 @@ ostream& operator<<(ostream& out, const vector<Node>& nodes) {
 }
 
 //=============================================================================
-using Distance = size_t;
+using Distance = double;
 constexpr Distance MAX_EDGE_DISTANCE = 100;
 constexpr Distance MAX_DISTANCE = NUM_NODES * MAX_EDGE_DISTANCE;
 class Edge {
